services/cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gridmind-cli"
version = "0.1.0"
description = "GridMind CLI — manage agent teams, deployments, costs, and auth."
requires-python = ">=3.12"
dependencies = [
    "click==8.1.7",
    "httpx==0.27.0",
    "pyyaml==6.0.1",
    "rich==13.7.1",
    "pydantic==2.6.1",
]

[project.scripts]
gridmind = "gridmind_cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
